feat: rename files to match classes names
