+++ src/aws/AWSConfigManager.ts
