+++ src/aws/index.ts
@@ -1,4 +1,4 @@
-export * from "./buildConfig.js";
+export * from "./AWSConfigManager.js";
 export * from "./DynamoDB.js";
 export * from "./ECS.js";
 export * from "./schema.js";
