+++ src/config/index.ts
@@ -1,3 +1,3 @@
-export { ConfigManager } from "./loadConfig.js";
+export { ConfigManager } from "./ConfigManager.js";
 export { parsePrefixedAddress } from "./parsePrefixedAddress.js";
 export { PrefixedAddress, Schema } from "./schema.js";
